requirements.txt
argon2-cffi==20.1.0
asgiref==3.3.1
cffi==1.14.4
Django==3.1.6
django-cors-headers==3.7.0
djangorestframework==3.12.2
psycopg2==2.8.6
pycparser==2.20
pytz==2021.1
six==1.15.0
sqlparse==0.4.1
